+++ sys/kern/vfs_bio.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: vfs_bio.c,v 1.178 2016/09/16 02:54:51 dlg Exp $	*/
+/*	$OpenBSD: vfs_bio.c,v 1.179 2017/02/14 10:31:15 mpi Exp $	*/
 /*	$NetBSD: vfs_bio.c,v 1.44 1996/06/11 11:15:36 pk Exp $	*/
 
 /*
@@ -893,7 +893,7 @@ start:
  * Get an empty, disassociated buffer of given size.
  */
 struct buf *
-geteblk(int size)
+geteblk(size_t size)
 {
 	struct buf *bp;
